+++ editor/src/components/editor/editor-component.tsx
@@ -73,6 +73,7 @@ import { isRoomId, projectIdToRoomId } from '../../utils/room-id'
 import { SharingDialog } from './sharing-dialog'
 import { AccessLevelParamKey, CloneParamKey } from './persistence/persistence-backend'
 import { useUpdateActiveRemixSceneOnSelectionChange } from '../canvas/remix/utopia-remix-root-component'
+import { useDefaultCollapsedViews } from './use-mama'
 
 const liveModeToastId = 'play-mode-toast'
 
@@ -423,6 +424,8 @@ export const EditorComponentInner = React.memo((props: EditorProps) => {
   )
 
   useLiveblocksConnectionListener()
+
+  useDefaultCollapsedViews()
 
   return (
     <>
