+++ editor/src/components/navigator/navigator.tsx
@@ -15,7 +15,6 @@ import Utils from '../../utils/utils'
 import { FlexColumn, Section, SectionBodyArea } from '../../uuiui'
 import { setFocus } from '../common/actions'
 import {
-  addCollapsedViews,
   clearHighlightedViews,
   clearSelection,
   showContextMenu,
@@ -32,8 +31,6 @@ import { ElementContextMenu } from '../element-context-menu'
 import { getItemHeight } from './navigator-item/navigator-item'
 import { NavigatorDragLayer } from './navigator-drag-layer'
 import { NavigatorItemWrapper } from './navigator-item/navigator-item-wrapper'
-import { isRight } from '../../core/shared/either'
-import { isJSXElement } from '../../core/shared/element-template'
 
 interface ItemProps extends ListChildComponentProps {}
 
@@ -264,8 +261,6 @@ export const NavigatorComponent = React.memo(() => {
     [dispatch],
   )
 
-  useDefaultCollapsed()
-
   return (
     <Section
       data-name='Navigator'
@@ -328,53 +323,3 @@ export const NavigatorComponent = React.memo(() => {
   )
 })
 NavigatorComponent.displayName = 'NavigatorComponent'
-
-function useDefaultCollapsed() {
-  const dispatch = useDispatch()
-
-  // keep track of the previously collapsed element paths, or they will be re-collapsed
-  type PreviouslyCollapsed = { [path: string]: boolean }
-  const [previouslyCollapsed, setPreviouslyCollapsed] = React.useState<PreviouslyCollapsed>({})
-
-  const elementNamesByElementPath = useEditorState(
-    Substores.metadata,
-    (store) => {
-      const elements = Object.values(store.editor.jsxMetadata)
-      return elements.reduce((acc, element) => {
-        if (isRight(element.element) && isJSXElement(element.element.value)) {
-          acc[EP.toString(element.elementPath)] = element.element.value.name.baseVariable
-        }
-        return acc
-      }, {} as { [key: string]: string })
-    },
-    'useDefaultCollapsed elementNamesByElementPath',
-  )
-
-  const collapsedPaths = useEditorState(
-    Substores.navigator,
-    (store): Set<string> => new Set(store.editor.navigator.collapsedViews.map(EP.toString)),
-    'useDefaultCollapsed collapsedPaths',
-  )
-
-  React.useEffect(() => {
-    const newCollapsedViews = Object.keys(elementNamesByElementPath).filter((path) => {
-      return (
-        // the collapsed views don't already contain the element
-        !collapsedPaths.has(path) &&
-        // the collapsed view wasn't already collapsed before
-        previouslyCollapsed[path] == null &&
-        // the element is a `head`
-        elementNamesByElementPath[path] === 'head'
-      )
-    })
-
-    if (newCollapsedViews.length > 0) {
-      // 1. add the new collapsed views
-      dispatch([addCollapsedViews(newCollapsedViews.map(EP.fromString))])
-      // 2. store the new added views to the previous ones
-      setPreviouslyCollapsed(
-        newCollapsedViews.reduce((acc, path) => ({ ...acc, [path]: true }), previouslyCollapsed),
-      )
-    }
-  }, [elementNamesByElementPath, collapsedPaths, dispatch, previouslyCollapsed])
-}
